backend/requirements.txt
fastapi==0.115.2
uvicorn[standard]==0.30.6
pydantic==2.9.2
sqlalchemy==2.0.36
psycopg2-binary==2.9.9
duckdb==1.1.3
pandas==2.2.3
pyarrow==17.0.0
python-multipart==0.0.9
APScheduler==3.10.4
redis==5.0.8
httpx==0.27.2
passlib[bcrypt]==1.7.4
python-jose==3.3.0
email-validator==2.2.0
loguru==0.7.2
